order_queue/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-semantic-conventions==0.45b0